Changed most fns to int16_t type. Added more fns to wsa_api.cpp

+++ wsa4k_cli.h
@@ -15,7 +15,7 @@ extern uint8_t debug_mode;
 extern uint8_t test_mode;
 
 
-int32_t start_cli(void);
+int16_t start_cli(void);
 
 
 //void print_scpi_menu();
